+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [2.0.0a2] - 2020-04-27
+
+Pypi: https://pypi.org/project/cobib/2.0.0a2/
+
 ### Added
 - user configuration options for:
     - TUI colors